feat: integrate MyBatis-Plus and connect ProductService to MySQL
- Add mybatis-plus-spring-boot4-starter 3.5.17 and mysql-connector-j 9.7.0
- Upgrade Product.price from Double to BigDecimal for DECIMAL(10,2) accuracy
- Create ProductMapper extending BaseMapper<Product>
- Replace hardcoded List.of(...) in ProductServiceImpl with mapper.selectList(LambdaQueryWrapper)
- Decouple datasource config via spring.profiles.group, moved to application-local.yml (gitignored)

+++ src/main/java/com/okeychou/aishop/entity/Product.java
@@ -1,12 +1,14 @@
 package com.okeychou.aishop.entity;
 
+import java.math.BigDecimal;
+
 public class Product {
     private Long id;
     private String name;
     private String category;
-    private Double price;
+    private BigDecimal price;
     public Product() {}
-    public Product(Long id, String name, String category, Double price) {
+    public Product(Long id, String name, String category, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.category = category;
@@ -37,11 +39,11 @@ public class Product {
         this.category = category;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
