+++ src/main/java/com/okeychou/aishop/service/ProductService.java
@@ -2,8 +2,9 @@ package com.okeychou.aishop.service;
 
 import com.okeychou.aishop.entity.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductService {
-    List<Product> listByPriceRange(double min, double max);
+    List<Product> listByPriceRange(BigDecimal min, BigDecimal max);
 }
